created service and controller layers

+++ src/entity/Product.java
@@ -3,9 +3,9 @@ package entity;
 public abstract class Product {
     private Long id;
     private  String name;
-    private  Integer price;
+    private  Double price;
 
-    public Product(Long id, String name, Integer price) {
+    public Product(Long id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,11 +30,11 @@ public abstract class Product {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
